kafka_training/scripts/pyproject.toml
[tool.poetry]
name = "kafka-training"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.9"
invoke = "~1.7"
kafka-python = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
